Make UDir factory method and use it.
createUDir is the factory method.
Previous createUDir now becomes createUDriFromAbs.

Currently, we assume path separator is "/", but in windows it is not.
But I want to keep the same value for UPath and UDir, that is, "/" separated.
In this way the file under .uit/dirs/ and .uit/hash/ is the same for any platform.

So I will fix to replace backslash to slash for windows environment in the future.
This change hope reducing that work.